crowd: Add initial files
This adds the crowd module for integrating ReviewBoard with Atlassian Crowd.
This currently just acts as an additional authentication mechanism for ReviewBoard,
without actually acting as a ReviewBoard extension. This is so the module supports
versions of ReviewBoard prior to version 2.0.

This module was written by Joshua Colp <jcolp AT digium DOT com>